requirements.txt
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
backports.zoneinfo==0.2.1
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
inflect==7.0.0
kafka-python==1.4.3
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyMySQL==1.1.0
pytz==2024.1
six==1.16.0
sniffio==1.3.1
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.51
starlette==0.36.3
typing_extensions==4.10.0
tzlocal==5.2
uvicorn==0.27.1
